Update Akan translations
This is an update to mozilla-central (not the latest but close to it).
This also includes a change in DTD formatting, i.e. preserving spaces
between the entity name and value.  The translation count in 83%
translated.

Complete PO count data for this commit is:
TOTAL:
type              strings      words (source)    words (translation)
translated:    4894 ( 83%)      22033 ( 79%)           24482
fuzzy:          205 (  3%)        852 (  3%)             n/a
untranslated:   735 ( 12%)       4790 ( 17%)             n/a
Total:         5834             27675                  24482

File count:     228